chore: 전체적으로 지저분한 부분 제거, 주석 작성 및 코드 구분
