+++ SamsungStoreApp/SceneDelegate.swift
@@ -16,8 +16,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
         guard let windowScene = (scene as? UIWindowScene) else { return }
         let window = UIWindow(windowScene: windowScene)
-        //window.rootViewController = ViewController()
-      window.rootViewController = MenuCollectionViewController()
+        window.rootViewController = ViewController()
         window.makeKeyAndVisible()
         self.window = window
     }
