+++ SamsungStoreApp/View/MenuCollectionView/ProductGridCell.swift
@@ -8,18 +8,24 @@
 import SnapKit
 import UIKit
 
-class ProductGridCell: UIView {
+final class ProductGridCell: UIView {
+  // MARK: - UI 컴포넌트
+  
   let imageView = UIImageView()
   let nameLabel = UILabel()
   let priceLabel = UILabel()
+  
+  // MARK: - 모델
   
   var product: ProductItem? {
     didSet {
       imageView.image = UIImage(named: product?.image ?? "placeholder")
       nameLabel.text = product?.name ?? "-"
-      priceLabel.text = (product?.price ?? "0") + "원"
+      priceLabel.text = "\(product?.price ?? "0")원"
     }
   }
+  
+  // MARK: - 초기화
   
   override init(frame: CGRect) {
     super.init(frame: frame)
@@ -31,6 +37,8 @@ class ProductGridCell: UIView {
     fatalError("init(coder:) has not been implemented")
   }
   
+  // MARK: - UI 세팅
+  
   private func setupUI() {
     backgroundColor = .white
     layer.cornerRadius = 8
@@ -39,12 +47,12 @@ class ProductGridCell: UIView {
     clipsToBounds = true
     
     imageView.contentMode = .scaleAspectFit
-//    imageView.backgroundColor = .lightGray
     imageView.layer.cornerRadius = 8
     imageView.clipsToBounds = true
     
     nameLabel.font = Font.title(size: 14)
     nameLabel.textAlignment = .center
+    
     priceLabel.font = Font.text(size: 13)
     priceLabel.textColor = .secondaryLabel
     priceLabel.textAlignment = .center
@@ -52,13 +60,10 @@ class ProductGridCell: UIView {
     let stack = UIStackView(arrangedSubviews: [imageView, nameLabel, priceLabel])
     stack.axis = .vertical
     stack.spacing = 6
-    addSubview(stack)
     
+    addSubview(stack)
     stack.snp.makeConstraints {
-      $0.top.equalToSuperview().inset(8)
-      $0.bottom.equalToSuperview().inset(8)
-      $0.leading.equalToSuperview().inset(8)
-      $0.trailing.equalToSuperview().inset(8)
+      $0.edges.equalToSuperview().inset(8)
     }
     
     imageView.snp.makeConstraints {
