+++ SamsungStoreApp/View/MenuCollectionView/ProductPageCell.swift
@@ -9,17 +9,20 @@ import SnapKit
 import UIKit
 
 final class ProductPageCell: UICollectionViewCell {
-  // MARK: -
+  // MARK: - UI 컴포넌트
 
   private let verticalStack = UIStackView() // 수직 스택
   private var cells: [ProductGridCell] = []
+  
+  // MARK: - 핸들러
+  
   var tapHandler: ((ProductItem) -> Void)?
 
-  // MARK: -
+  // MARK: - 초기화
 
   override init(frame: CGRect) {
     super.init(frame: frame)
-    setupView()
+    setupUI()
     setupLayout()
   }
 
@@ -28,15 +31,14 @@ final class ProductPageCell: UICollectionViewCell {
     fatalError("init(coder:) has not been implemented")
   }
 
-  // MARK: -
+  // MARK: - UI 세팅
 
-  private func setupView() {
+  private func setupUI() {
     verticalStack.axis = .vertical
     verticalStack.spacing = 24
     verticalStack.distribution = .fillEqually
-    contentView.addSubview(verticalStack) // 컨텐츠 뷰에 수직 스택을 넣음. 그 수직 스택에는 수평 스택이 들어 있다.
+    contentView.addSubview(verticalStack)
 
-    // 2 rows
     for _ in 0..<2 {
       let rowStack = UIStackView()
       rowStack.axis = .horizontal
@@ -44,10 +46,10 @@ final class ProductPageCell: UICollectionViewCell {
       rowStack.distribution = .fillEqually
 
       for _ in 0..<2 {
-        let cell = ProductGridCell()
-        cell.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(cellTapped(_:))))
-        cells.append(cell)
-        rowStack.addArrangedSubview(cell)
+        let grid = ProductGridCell()
+        grid.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(cellTapped(_:))))
+        rowStack.addArrangedSubview(grid)
+        cells.append(grid)
       }
       verticalStack.addArrangedSubview(rowStack)
     }
@@ -59,7 +61,7 @@ final class ProductPageCell: UICollectionViewCell {
     }
   }
 
-  // MARK: -
+  // MARK: - 구성
   
   func configure(with products: [ProductItem]) {
     for (index, cell) in cells.enumerated() {
@@ -92,6 +94,8 @@ final class ProductPageCell: UICollectionViewCell {
         }
       }
   }
+  
+  // MARK: - 액션
 
   @objc private func cellTapped(_ sender: UITapGestureRecognizer) {
     guard let cell = sender.view as? ProductGridCell,
